6018-Create-Binary-Tree/Solution_test.go
package leetcode

import (
	"github.com/duanbiaowu/leetcode/structures"
	"reflect"
	"testing"
)

func Test_createBinaryTree(t *testing.T) {
	type args struct {
		descriptions [][]int
	}
	tests := []struct {
		name string
		args args
		want *TreeNode
	}{
		{
			"test-1",
			args{nil},
			nil,
		},
		{
			"test-2",
			args{[][]int{
				{1, 2, 1},
				{2, 3, 0},
				{3, 4, 1},
			}},
			structures.GenerateTreeNodesBySlice([]int{1, 2, structures.NULL, structures.NULL, 3, 4}),
		},
		{
			"test-3",
			args{[][]int{
				{20, 15, 1},
				{20, 17, 0},
				{50, 20, 1},
				{50, 80, 0},
				{80, 19, 1},
			}},
			structures.GenerateTreeNodesBySlice([]int{50, 20, 80, 15, 17, 19}),
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := createBinaryTree(tt.args.descriptions); !reflect.DeepEqual(got, tt.want) {
				t.Errorf("createBinaryTree() = %v, want %v", got, tt.want)
			}
		})
	}
}
